+++ docs/changes.rst
@@ -17,3 +17,5 @@ v0.1.0, ????-??-??
   <https://github.com/gradha/lazy_rest/issues/8>`_.
 * `Changed safe_* procs to return info when something went wrong
   <https://github.com/gradha/lazy_rest/issues/4>`_.
+* `Added API to customize safe error message
+  <https://github.com/gradha/lazy_rest/issues/15>`_.